+++ src/Kernel.php
@@ -141,9 +141,13 @@ class Kernel extends BaseKernel implements CompilerPassInterface
 
     protected function configureRoutes(RouteCollectionBuilder $routes): void
     {
+        $routes->import("{$this->resourceDir}/routes/{$this->environment}/*.yml", '/', 'glob');
+        $routes->import("{$this->resourceDir}/{routes}/*.yml", '/', 'glob');
+
         $routes->import("{$this->resourceDir}/routes_reports.yml");
         $routes->import("{$this->resourceDir}/routes_manage.yml");
         $routes->import("{$this->resourceDir}/routes.yml");
+
         // optional routes by local install
         if (file_exists("{$this->configDir}/routes.yml")) {
             $routes->import("{$this->configDir}/routes.yml");