+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 See [Upgrading] for details how to upgrade.
 
 - Drop loading deprecated `ldap.php`, #899
+- Update `laminas/laminas-mail` (2.11.0 => 2.12.0), #896
 
 [3.9.3]: https://github.com/eventum/eventum/compare/v3.9.2...master